refactor(java): isolate topology types
