+++ src/Java/myservicebus-rabbitmq-tests/src/test/java/com/myservicebus/rabbitmq/RabbitMqFactoryConfiguratorTests.java
@@ -7,6 +7,7 @@ import java.util.concurrent.CompletableFuture;
 import org.junit.jupiter.api.Test;
 
 import com.myservicebus.*;
+import com.myservicebus.topology.*;
 import com.myservicebus.di.ServiceCollection;
 import com.myservicebus.di.ServiceProvider;
