+++ src/Java/myservicebus-rabbitmq/src/main/java/com/myservicebus/RabbitMqMessageBus.java
@@ -20,6 +20,9 @@ import com.myservicebus.di.ServiceCollection;
 import com.myservicebus.di.ServiceProvider;
 import com.myservicebus.tasks.CancellationToken;
 import com.myservicebus.rabbitmq.ConnectionProvider;
+import com.myservicebus.topology.ConsumerTopology;
+import com.myservicebus.topology.MessageBinding;
+import com.myservicebus.topology.TopologyRegistry;
 import com.rabbitmq.client.BuiltinExchangeType;
 import com.rabbitmq.client.Channel;
 import com.rabbitmq.client.Connection;
