+++ src/Java/myservicebus/src/main/java/com/myservicebus/topology/TopologyRegistry.java
@@ -1,8 +1,12 @@
-package com.myservicebus;
+package com.myservicebus.topology;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Consumer;
+
+import com.myservicebus.ConsumeContext;
+import com.myservicebus.NamingConventions;
+import com.myservicebus.PipeConfigurator;
 
 public class TopologyRegistry {
     private final List<MessageTopology> messages = new ArrayList<>();
